refactor: move isSocialCardString to utils-common
